@@ -1,16 +1,14 @@
 module github.com/kevwan/tproxy
 
-go 1.16
+go 1.18
 
 require (
 	github.com/fatih/color v1.13.0
 	github.com/grpc/grpc/tools/http2_interop v0.0.0-20220615001509-d0751191a921
-	github.com/olekukonko/tablewriter v0.0.5
 )
 
 require (
 	github.com/mattn/go-colorable v0.1.9 // indirect
 	github.com/mattn/go-isatty v0.0.14 // indirect
-	github.com/mattn/go-runewidth v0.0.9 // indirect
 	golang.org/x/sys v0.0.0-20210630005230-0f9fa26af87c // indirect
 )
